@@ -7,6 +7,9 @@
 #include <utility>
 #include <vector>
 
+#include "M2FreeAlgebra.hpp"
+#include "M2FreeAlgebraQuotient.hpp"
+
 #include "aring-CC.hpp"
 #include "aring-CCC.hpp"
 #include "aring-glue.hpp"
@@ -401,26 +404,36 @@ const RingElement /* or null */ *IM2_RingElement_term(const Ring *R,
    coefficient ring of R.  Returns a*m, if this is a valid
    element of R.  Returns NULL if not (with an error message). */
 {
-  try
-    {
-      const PolynomialRing *P = R->cast_to_PolynomialRing();
-      if (P == 0)
-        {
-          ERROR("requires a polynomial ring");
-          return 0;
-        }
-
-      int nvars0 = P->n_vars();
-      const PolynomialRing *K = a->get_ring()->cast_to_PolynomialRing();
-      if (K != 0 && K != P->getCoefficients()) nvars0 -= K->n_vars();
-      int *exp = newarray_atomic(int, nvars0);
-      varpower::to_ntuple(nvars0, m->ints(), exp);
-      ring_elem val = P->make_logical_term(a->get_ring(), a->get_value(), exp);
-      return RingElement::make_raw(R, val);
-  } catch (const exc::engine_error& e)
-    {
-      ERROR(e.what());
-      return NULL;
+  try {
+    const PolynomialRing *P = R->cast_to_PolynomialRing();
+    if (P != nullptr)
+      {
+        int nvars0 = P->n_vars();
+        const PolynomialRing *K = a->get_ring()->cast_to_PolynomialRing();
+        if (K != nullptr && K != P->getCoefficients()) nvars0 -= K->n_vars();
+        int *exp = newarray_atomic(int,nvars0);
+        varpower::to_ntuple(nvars0, m->ints(), exp);
+        ring_elem val = P->make_logical_term(a->get_ring(), a->get_value(), exp);
+        return RingElement::make_raw(R,val);
+      }
+    auto Q = dynamic_cast<const M2FreeAlgebraOrQuotient *>(R);
+    if (Q != nullptr)
+      {
+        if (Q->coefficientRing() != a->get_ring())
+          {
+            ERROR("wrong coefficient ring");
+            return nullptr;
+          }
+        return RingElement::make_raw(Q,
+                                     Q->makeTerm(a->get_value(),
+                                                 m->ints()));
+      }
+    ERROR("requires a polynomial ring");
+    return nullptr;
+  }
+  catch (exc::engine_error& e) {
+    ERROR(e.what());
+    return nullptr;
   }
 }
 
@@ -495,24 +508,35 @@ engine_RawArrayPairOrNull IM2_RingElement_list_form(
     const Ring *coeffRing, /* ring of the result coefficients */
     const RingElement *f)
 {
-  try
-    {
-      const PolynomialRing *P = f->get_ring()->cast_to_PolynomialRing();
-      if (P != 0)
-        {
-          return P->list_form(coeffRing, f->get_value());
-        }
-      const SchurRing2 *S = f->get_ring()->cast_to_SchurRing2();
-      if (S != 0)
-        {
-          return S->list_form(coeffRing, f->get_value());
-        }
-      ERROR("expected a polynomial");
-      return 0;
-  } catch (const exc::engine_error& e)
-    {
-      ERROR(e.what());
-      return NULL;
+  try {
+    const PolynomialRing* P = f->get_ring()->cast_to_PolynomialRing();
+    if (P != nullptr)
+      {
+        return P->list_form(coeffRing, f->get_value());
+      }
+    const SchurRing2* S = f->get_ring()->cast_to_SchurRing2();
+    if (S != nullptr)
+      {
+        return S->list_form(coeffRing, f->get_value());
+      }
+    /* added by Frank */
+    const M2FreeAlgebra* ncP = dynamic_cast<const M2FreeAlgebra*>(f->get_ring());
+    if (ncP != nullptr)
+      {
+        return ncP->list_form(coeffRing, f->get_value());
+      }
+    /* added by Frank */
+    const M2FreeAlgebraQuotient* ncQ = dynamic_cast<const M2FreeAlgebraQuotient*>(f->get_ring());
+    if (ncQ != nullptr)
+      {
+        return ncQ->m2FreeAlgebra().list_form(coeffRing, f->get_value());
+      }
+    ERROR("expected a polynomial");
+    return nullptr;
+  }
+  catch (exc::engine_error& e) {
+    ERROR(e.what());
+    return nullptr;
   }
 }
 
