@@ -117,7 +117,11 @@ void GaloisFieldTable::display(std::ostream &o) const
   o << std::endl;
 }
 
-ARingGFM2::ARingGFM2(const PolynomialRing &R, const ring_elem a) : mGF(R, a) {}
+ARingGFM2::ARingGFM2(const PolynomialRing &R, const ring_elem a) : mGF(R, a)
+{
+  // Nothing to do here.
+}
+
 void ARingGFM2::fromSmallIntegerCoefficients(
     ElementType &result,
     const std::vector<long> &poly) const
@@ -196,7 +200,7 @@ void ARingGFM2::elem_text_out(buffer &o,
   mGF.ring().elem_text_out(o, h, p_one, p_plus, p_parens);
   mGF.ring().remove(h);
 }
-};
+}; // namespace M2
 
 // Local Variables:
 // compile-command: "make -C $M2BUILDDIR/Macaulay2/e "
