@@ -17,8 +17,9 @@
 
 // this fixes a weird problem in the package "openblas" of Arch Linux, which somehow fails to declare this function:
 extern "C" void openblas_set_num_threads(int num_threads);
-
+#define bool_constant givaro_bool_constant
 #include <fflas-ffpack/ffpack/ffpack.h>
+#undef bool_constant
 #pragma GCC diagnostic pop
 
 namespace M2 {
