@@ -17,13 +17,15 @@
 #include "weylalg.hpp"
 #include "skewpoly.hpp"
 #include "solvable.hpp"
-#include "polyquotient.hpp"
-
 #include "matrix.hpp"
 #include "exceptions.hpp"
 #include "finalize.hpp"
-
 #include "tower.hpp"
+
+#include "Polynomial.hpp"
+#include "M2FreeAlgebra.hpp"
+#include "M2FreeAlgebraQuotient.hpp"
+#include "polyquotient.hpp"
 
 #include "aring.hpp"
 #include "aring-glue.hpp"
@@ -205,6 +207,93 @@ const Ring /* or null */ *IM2_Ring_solvable_algebra(const Ring *R,
   }
 }
 
+const Ring* /* or null */ rawRingM2FreeAlgebra(const Ring* coefficientRing,
+                                               M2_ArrayString names,
+                                               const Ring* degreeRing,
+                                               M2_arrayint degrees,
+                                               M2_arrayint wtvecs,
+                                               M2_arrayint heftVector)
+{
+  try {
+    if (coefficientRing == nullptr)
+      {
+        ERROR("internal error: expected non-null Ring!");
+        return nullptr;
+      }
+    const PolynomialRing *P = degreeRing->cast_to_PolynomialRing();
+    if (P == nullptr)
+      {
+        ERROR("expected polynomial ring");
+        return nullptr;
+      }
+    const M2FreeAlgebra* result = M2FreeAlgebra::create(coefficientRing,
+                                                        M2_ArrayString_to_stdvector(names),
+                                                        P,
+                                                        M2_arrayint_to_stdvector<int>(degrees),
+                                                        M2_arrayint_to_stdvector<int>(wtvecs),
+                                                        M2_arrayint_to_stdvector<int>(heftVector));
+    //intern_polyring(result); // we might want to intern our rings (to register a finalizer with the gc)
+    return result;
+  }
+  catch (exc::engine_error& e) {
+    ERROR(e.what());
+    return NULL;
+  }
+}
+
+/* WIP
+const M2FreeMonoid* rawM2FreeMonoid(M2_ArrayString names,
+                                    const Ring* degreeRing,
+                                    M2_arrayint degrees,
+                                    M2_arrayint wtvecs,
+                                    M2_arrayint heftVector)
+{
+  try {
+    const PolynomialRing *P = degreeRing->cast_to_PolynomialRing();
+    if (P == nullptr)
+      {
+        ERROR("expected polynomial ring");
+        return nullptr;
+      }
+    const M2FreeMonoid* result = M2FreeMonoid::create(M2_ArrayString_to_stdvector(names),
+                                                      P,
+                                                      M2_arrayint_to_stdvector<int>(degrees),
+                                                      M2_arrayint_to_stdvector<int>(wtvecs),
+                                                      M2_arrayint_to_stdvector<int>(heftVector));
+    return result;
+  }
+  catch (exc::engine_error& e) {
+    ERROR(e.what());
+    return NULL;
+  }
+}
+*/
+
+const Ring* /* or null */ rawRingM2FreeAlgebraQuotient(const Matrix* GB, int maxdeg)
+{
+  const Ring* A = GB->get_ring();
+  try {
+    if (A == nullptr)
+      {
+        ERROR("internal error: expected non-null Ring!");
+        return nullptr;
+      }
+    const M2FreeAlgebra* P = A->cast_to_M2FreeAlgebra();
+    if (P == nullptr)
+      {
+        ERROR("expected a free algebra");
+        return nullptr;
+      }
+    
+    const M2FreeAlgebraQuotient* result = M2FreeAlgebraQuotient::create(* P, GB, maxdeg);
+    return result;
+  }
+  catch (exc::engine_error& e) {
+    ERROR(e.what());
+    return NULL;
+  }
+}
+
 const Ring /* or null */ *IM2_Ring_frac(const Ring *R)
 {
   try
@@ -297,14 +386,20 @@ const Ring /* or null */ *IM2_Ring_quotient(const Ring *R, const Matrix *I)
           return 0;
         }
       const PolynomialRing *P = R->cast_to_PolynomialRing();
-      if (P == 0)
+      if (P != nullptr)
         {
-          ERROR("expected a polynomial ring");
-          return 0;
+          PolynomialRing *result = PolynomialRing::create_quotient(P, I);
+          intern_polyring(result);
+          return result;
         }
-      PolynomialRing *result = PolynomialRing::create_quotient(P, I);
-      intern_polyring(result);
-      return result;
+      const M2FreeAlgebra *A = R->cast_to_M2FreeAlgebra();
+      if (A != nullptr)
+        {
+          auto result = M2FreeAlgebraQuotient::create(*A, I, -1);
+          return result;
+        }
+      ERROR("expected a polynomial ring or free algebra");
+      return nullptr;
   } catch (const exc::engine_error& e)
     {
       ERROR(e.what());
