@@ -6,6 +6,7 @@
 #include "frac.hpp"
 #include "localring.hpp"
 #include "polyring.hpp"
+#include "M2FreeAlgebra.hpp"
 
 #include "aring-glue.hpp"
 
@@ -24,12 +25,18 @@ RingElement *RingElement::make_raw(const Ring *R, ring_elem f)
 int RingElement::n_terms(int nvars) const
 {
   const PolynomialRing *P = R->cast_to_PolynomialRing();
-  if (P == 0)
+  if (is_zero()) return 0;
+  if (P != nullptr)
     {
-      if (is_zero()) return 0;
-      return 1;
+      return P->n_logical_terms(nvars, val);
     }
-  return P->n_logical_terms(nvars, val);
+  auto Q = dynamic_cast<const M2FreeAlgebra *>(R);
+  if (Q != nullptr)
+    {
+      return Q->n_terms(val);
+    }
+  
+  return 1;
 }
 
 RingElement *RingElement::operator-() const
@@ -132,33 +139,47 @@ RingElement *RingElement::random(const Ring *R)
   return new RingElement(R, R->random());
 }
 
-void RingElement::text_out(buffer &o) const { R->elem_text_out(o, val); }
+void RingElement::text_out(buffer &o) const
+{
+  R->elem_text_out(o, val);
+}
+
 RingElement /* or null */ *RingElement::get_terms(int nvars,
                                                   int lo,
                                                   int hi) const
 {
   const PolynomialRing *P = R->cast_to_PolynomialRing();
-  if (P == 0)
+  if (P != nullptr)
     {
-      ERROR("expected polynomial ring");
-      return 0;
+      return new RingElement(P, P->get_terms(nvars, val, lo, hi));
     }
-  return new RingElement(P, P->get_terms(nvars, val, lo, hi));
+  const M2FreeAlgebra* A = dynamic_cast<const M2FreeAlgebra*>(R);
+  if (A != nullptr)
+    {
+      return new RingElement(A, A->get_terms(val, lo, hi));
+    }
+  ERROR("expected polynomial ring");
+  return nullptr;
 }
 
 RingElement /* or null */ *RingElement::lead_coeff(const Ring *coeffR) const
 {
-  const PolynomialRing *P = R->cast_to_PolynomialRing();
-  if (P == 0)
-    {
-      ERROR("expected polynomial ring");
-      return 0;
-    }
   if (is_zero())
     {
       return new RingElement(coeffR, coeffR->zero());
     }
-  return new RingElement(coeffR, P->lead_logical_coeff(coeffR, val));
+  const PolynomialRing *P = R->cast_to_PolynomialRing();
+  if (P != nullptr)
+    {
+      return new RingElement(coeffR, P->lead_logical_coeff(coeffR, val));
+    }
+  const M2FreeAlgebra* A = dynamic_cast<const M2FreeAlgebra*>(R);
+  if (A != nullptr)
+    {
+      return new RingElement(coeffR, A->lead_coefficient(coeffR, val));
+    }
+  ERROR("expected polynomial ring");
+  return nullptr;
 }
 
 RingElement /* or null */ *RingElement::get_coeff(const Ring *coeffR,
@@ -175,25 +196,30 @@ RingElement /* or null */ *RingElement::get_coeff(const Ring *coeffR,
 
 Monomial *RingElement::lead_monom(int nvars) const
 {
-  const PolynomialRing *P = R->cast_to_PolynomialRing();
-  if (P == 0)
-    {
-      ERROR("expected polynomial ring");
-      return 0;
-    }
   if (is_zero())
     {
-      ERROR("zero polynomial has no lead monomial");
-      return 0;
+      ERROR("the zero element has no lead monomial");
+      return nullptr;
     }
+  const PolynomialRing *P = R->cast_to_PolynomialRing();
+  if (P != nullptr)
+    {
+      intarray resultvp;
+      Nterm *t = get_value();
 
-  intarray resultvp;
-  Nterm *t = get_value();
-
-  int *exp = newarray_atomic(int, nvars);
-  P->lead_logical_exponents(nvars, t, exp);
-  varpower::from_ntuple(nvars, exp, resultvp);
-  return Monomial::make(resultvp.raw());
+      int *exp = newarray_atomic(int, nvars);
+      P->lead_logical_exponents(nvars, t, exp);
+      varpower::from_ntuple(nvars, exp, resultvp);
+      return Monomial::make(resultvp.raw());
+    }
+  const M2FreeAlgebraOrQuotient* Q = dynamic_cast<const M2FreeAlgebraOrQuotient*>(R);
+  if (Q != nullptr)
+    {
+      ERROR("not implemented yet");
+      return nullptr;
+    }
+  ERROR("expected polynomial ring");
+  return nullptr;
 }
 
 bool RingElement::is_homogeneous() const { return R->is_homogeneous(val); }
