@@ -10,6 +10,7 @@
 
 #include "RingTest.hpp"
 #include "tower.hpp"
+#include "../util.hpp"
 
 // First: we need a routne to read a polynomial from a string.
 // Format:  variables are a..zA..Z, and then [1], [2], ...
@@ -18,17 +19,6 @@
 // coefficients: (+ or - or nothing) (number) (optional: . or /, followed by
 // another (number)
 // for GF, do we mix the a^r in?
-
-M2_ArrayString toM2ArrayString(std::vector<std::string>& strs)
-{
-  int n = static_cast<int>(
-      strs.size());  // needed since M2_ArrayString len field is int
-  int i;
-  M2_ArrayString a = getmemarraytype(M2_ArrayString, n);
-  a->len = n;
-  for (i = 0; i < n; i++) a->array[i] = M2_tostring(strs[i].c_str());
-  return a;
-}
 
 template <>
 ring_elem getElement<Tower>(const Tower& R, int index)
