@@ -17,8 +17,8 @@ class varpower
   static int compare(const int *a, const int *b);
   // return EQ, LT, or GT for a == b, a < b, or a > b.
 
-  // [len, v1, e1, ..., vn, en], where len=n or 2*n (which??)
- public:
+  // [len, v1, e1, ..., vn, en], where len=n
+public:
   varpower() {}
   ~varpower() {}
   static unsigned int computeHashValue(const int *vp);
